requirements.txt
transformers>=4.40.0
datasets>=2.19.0
torch>=2.1.0
huggingface_hub>=0.22.0
evaluate>=0.4.1
rouge-score>=0.1.2
sacrebleu>=2.4.0
click>=8.1.7
pyyaml>=6.0.1
tqdm>=4.66.0
pandas>=2.2.0
tabulate>=0.9.0
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.5.0
groq>=0.5.0
ollama>=0.1.8
pytest>=8.1.0
